feat: add service announcer command
This command is a Kubernetes controller that watches Services and
creates an associated l4proxy configuration file from the LoadBalancer
ones.

This controller is meant to be run either outside of Kubernetes so
that l4proxy can access the configuration or save the configuration
file to a volume mounted through e.g. NFS.