Make health interval configurable per service
The service-announcer now retrieves the annotation value for the
`l4proxy.e13.dev/health-interval` annotation from each service and, if
the annotation exists, sets the health interval on the respective
frontend in the l4proxy configuration.

This commit also gets rid of an unnecessary 2nd for loop when running
through all the services and instead directly creates the respective
l4proxy frontend configuration when iterating over all services.

closes #128